Replace filtered cohort matrix tests with class page matrix tests
