+++ pages/boac/api_admin_page.rb
@@ -5,7 +5,7 @@ class ApiAdminPage
   include PageObject
   include Logging
 
-  h1(:unauth_msg, xpath: '//h1[text()="Unauthorized"]')
+  h1(:unauth_msg, xpath: '//*[contains(.,"Unauthorized")]')
 
   def load_cachejob
     navigate_to "#{BOACUtils.base_url}/api/admin/cachejob"
