+++ util/spec_helper.rb
@@ -73,11 +73,11 @@ require_relative '../pages/boac/boac_pages'
 require_relative '../pages/boac/admin_page'
 require_relative '../pages/boac/home_page'
 require_relative '../pages/boac/cohort_pages'
-require_relative '../pages/boac/class_page'
+require_relative '../pages/boac/class_pages'
+require_relative '../pages/boac/class_list_view_page'
+require_relative '../pages/boac/class_matrix_view_page'
 require_relative '../pages/boac/curated_cohort_page'
-require_relative '../pages/boac/curated_cohort_list_view_page'
 require_relative '../pages/boac/filtered_cohort_page'
-require_relative '../pages/boac/filtered_cohort_list_view_page'
 require_relative '../pages/boac/search_results_page'
 require_relative '../pages/boac/student_page'
 require_relative '../pages/boac/teams_list_page'
